+++ src/system.go
@@ -1,6 +1,9 @@
 package blackboard_rest
 
 import (
+  "github.com/jbaxe2/blackboard.rest.go/src/_scaffolding/config"
+  error2 "github.com/jbaxe2/blackboard.rest.go/src/_scaffolding/error"
+  "github.com/jbaxe2/blackboard.rest.go/src/_scaffolding/factory"
   "github.com/jbaxe2/blackboard.rest.go/src/_scaffolding/services"
   "github.com/jbaxe2/blackboard.rest.go/src/oauth2"
   "github.com/jbaxe2/blackboard.rest.go/src/system"
@@ -49,7 +52,17 @@ func GetSystemInstance (
 func (restSystem *_BbRestSystem) GetVersion() (system.VersionInfo, error) {
   var version system.VersionInfo
   var err error
-  //var result interface{}
+  var result interface{}
+
+  result, err = restSystem.service.Connector.SendBbRequest (
+    config.SystemEndpoints["version"], "GET", make (map[string]interface{}), 1,
+  )
+
+  if (nil != err) && (error2.RestError{} != err) {
+    return version, err
+  }
+
+  version = factory.NewVersionInfo (result.(map[string]interface{}))
 
   return version, err
 }