+++ test/blackboard_rest_test.go
@@ -30,9 +30,12 @@ type BlackboardRestTester struct {
  * The [Run] method...
  */
 func (tester *BlackboardRestTester) Run() {
-  oauth2Tester := OAuth2Tester {t: tester.t}
-  oauth2Tester.Run()
+  //oauth2Tester := OAuth2Tester {t: tester.t}
+  //oauth2Tester.Run()
 
+  systemTester := SystemTester {t: tester.t}
+  systemTester.Run()
+/*
   usersTester := UsersTester {t: tester.t}
   usersTester.Run()
 
@@ -45,5 +48,8 @@ func (tester *BlackboardRestTester) Run() {
   termsTester := TermsTester {t: tester.t}
   termsTester.Run()
 
+  courseGradesTester := CourseGradesTester {t: tester.t}
+  courseGradesTester.Run()
+*/
   println()
 }